label templates and split out converter code into utilities
