+++ page-templates/guides-sidebar.html
@@ -3,18 +3,21 @@
   <div class="widget guide-list">
     <h6 class="montserrat text-uppercase bottom-line">Install</h6>
     <ul class="icons-list">
+      <!-- INSTALL-LIST -->
       <li><a href="install-linux.html">Install on Linux</a></li>
       <li><a href="install-mac.html">Install on Mac OS</a></li>
       <li><a href="install-windows.html">Install on Windows</a></li>
       <!--<li><a data-toggle="" href="">Quick Sync (Torrent) </a></li>-->
+      <!-- /INSTALL-LIST -->
     </ul>
     <br>
     <h6 class="montserrat text-uppercase bottom-line">Guides</h6>
     <ul class="icons-list">
+      <!-- GUIDES-LIST -->
       <li><a href="generate-address.html">Generate A Bitcoin Address</a></li>
       <li><a href="scripting.html">Intro to Scripting</a></li>
       <li><a href="op_return.html">Create an OP_RETURN</a></li>
-      <!--<li><a data-toggle="" href="">Quick Sync (Torrent) </a></li>-->
+      <!-- /GUIDES-LIST -->
     </ul>
   </div>
   <!-- END CATEGORIES WIDGET -->
