+++ page-templates/guides-wrapper.html
@@ -23,8 +23,8 @@
 
         <div class="row">
 
-          <!-- SIDEBAR -->
-          <!-- END SIDEBAR -->
+          <!-- GUIDES-SIDEBAR -->
+          <!-- /GUIDES-SIDEBAR -->
 
           <!-- START OF ARTICLE CONTAINER -->
           <div class="col-sm-9 blog-content post-thumbnail">
@@ -86,8 +86,10 @@
     </section>
     <!-- END PARALLAX DOCS CTA -->
 
-    <!-- FOOTER -->
-    <!-- END FOOTER -->
+    <!-- FOOTER-SECTION -->
+    <!-- /FOOTER-SECTION -->
 
   </div>
   <!-- /WRAPPER -->
+  <!-- FOOTER-SCRIPTS -->
+  <!-- /FOOTER-SCRIPTS -->
