+++ page-templates/page-wrapper.html
@@ -66,6 +66,9 @@
     </script>
 </head>
 <body>
-
+<!-- HEADER-SECTION -->
+<!-- /HEADER-SECTION -->
+<!-- GUIDES-WRAPPER -->
+<!-- /GUIDES-WRAPPER -->
 </body>
 </html>
